requirements.txt
# Web框架
Flask>=2.3.0
flask-cors>=4.0.0

# 任务队列（可选，支持MockCelery模式）
celery>=5.3.0

# AI/LLM相关
langchain>=0.0.350
langchain-community>=0.0.10

# HTTP客户端
requests>=2.31.0
aiohttp>=3.9.0

# 配置文件处理
PyYAML>=6.0.0
python-dotenv>=1.0.0

# 加密和安全
cryptography>=41.0.0

# Redis（可选，仅在USE_REDIS=true时需要）
redis>=5.0.0
